Eval: move eval_call_callable() to interpreter/utils
